+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 ## Changelog
 
-### 0.2.0 (Next)
+### 0.2.0 (11/21/2016)
 
 * [#13](https://github.com/dblock/danger-changelog/pull/13): Fix: suggested CHANGELOG entry may not be of correct format - [@dblock](https://github.com/dblock).
 * [#18](https://github.com/dblock/danger-changelog/pull/18): Fix: incorrect CHANGELOG filename reported when file doesn't exist - [@dblock](https://github.com/dblock).
-* Your contribution here.
 
 ### 0.1.0 (8/26/2016)